refactor(permissions): move user permissions in Session engine-side (#2126)
Closes [RUNTIME-638](https://bonitasoft.atlassian.net/browse/RUNTIME-638)

+++ bonita-integration-tests/bonita-integration-tests-client/src/main/java/org/bonitasoft/engine/business/application/TestWithLivingApplication.java
@@ -57,7 +57,9 @@ public class TestWithLivingApplication extends CommonAPIIT {
         final SearchResult<Application> searchResult = getLivingApplicationAPI()
                 .searchApplications(new SearchOptionsBuilder(0, 1000).done());
         for (final Application app : searchResult.getResult()) {
-            getLivingApplicationAPI().deleteApplication(app.getId());
+            if (app.isEditable()) {
+                getLivingApplicationAPI().deleteApplication(app.getId());
+            }
         }
         logoutThenlogin();
         deleteUser(user);
